Move changeset conventions into their own files and move things to appropriate locations
