+++ source/Calamari.Aws/Integration/CloudFormation/ChangeSetArn.cs
@@ -1,0 +1,12 @@
+﻿namespace Calamari.Aws.Integration.CloudFormation
+{
+    public class ChangeSetArn
+    {
+        public string Value { get; }
+
+        public ChangeSetArn(string value)
+        {
+            Value = value;
+        }
+    }
+}
