+++ source/Calamari.Aws/Integration/CloudFormation/StackArn.cs
@@ -1,0 +1,12 @@
+﻿namespace Calamari.Aws.Integration.CloudFormation
+{
+    public class StackArn
+    {
+        public StackArn(string value)
+        {
+            Value = value;
+        }
+
+        public string Value { get; }
+    }
+}
