+++ source/Calamari.Aws/Integration/CloudFormation/Templates/StackFormationNamedOutput.cs
@@ -1,0 +1,12 @@
+﻿namespace Calamari.Aws.Integration.CloudFormation.Templates
+{
+    public class StackFormationNamedOutput
+    {
+        public string Name { get; }
+
+        public StackFormationNamedOutput(string name)
+        {
+            Name = name;
+        }
+    }
+}
