+++ source/Calamari.Aws/Integration/S3/S3UploadResult.cs
@@ -1,7 +1,7 @@
 using Amazon.S3.Model;
 using Octopus.CoreUtilities;
 
-namespace Calamari.Aws.Deployment.Conventions
+namespace Calamari.Aws.Integration.S3
 {
     public class S3UploadResult
     {
